build(workflows): Extract git_functions to separate file, optimize imports
See: #28
