+++ src/visualization_functions.py
@@ -1,31 +1,10 @@
-import os
-import random
-import string
-import subprocess
+from datetime import datetime
+from pathlib import Path
+from random import choice
 
 import matplotlib.pyplot as plt
-from datetime import datetime
-from random import choice
-from pathlib import Path
-from PIL import Image
 import numpy as np
-
-
-# Function to fetch the current git branch name
-def get_git_branch():
-    github_ref = os.getenv('GITHUB_REF')
-    if github_ref and github_ref.startswith('refs/heads/'):
-        return github_ref.replace('refs/heads/', '')
-
-    try:
-        branch = subprocess.check_output(["git", "rev-parse", "--abbrev-ref", "HEAD"]).strip().decode()
-        return branch
-    except subprocess.CalledProcessError:
-        return "unknown-branch"
-
-
-def generate_short_hash():
-    return ''.join(random.choices(string.ascii_lowercase + string.digits, k=6))
+from PIL import Image
 
 
 def show_n_samples(dataset: dict, species_folders: dict, n_of_images: int = 5):
